requirements.txt
openai
python_dotenv
streamlit